% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age.R
\name{add_age}
\alias{add_age}
\title{Add Age to Data Frame}
\usage{
add_age(df)
}
\arguments{
\item{df}{A data frame containing the columns `study`, `histology_assessment_date`,
`birth_date_pat`, and `year_of_birth`. The `study` column is used to determine
the method of age calculation. The `histology_assessment_date` must be in a
date format that `lubridate::year()` can recognize.}
}
\value{
Returns the input data frame with an additional column `age` that contains
  the calculated age of the subjects.
}
\description{
This function calculates the age of subjects in a data frame based on their
year of birth or birth date and the date of histology assessment. The method
of calculation depends on the study specified in the data frame.
}
\examples{
df <- data.frame(
  study = c("embrace_i", "other_study"),
  histology_assesment_date = as.Date(c("2024-01-01", "2024-01-02")),
  birth_date_pat = c(1980, NA),
  year_of_birth = c(NA, 1990)
)
df_with_age <- add_age(df)
}
