% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_failure_event.R
\name{add_local_failure_event}
\alias{add_local_failure_event}
\title{Add a Column for Local Failure Events}
\usage{
add_local_failure_event(df)
}
\arguments{
\item{df}{A data frame containing at least one column starting with 'overall_disease_local_recurrence_'.
These columns should contain numerical values where 2 indicates a local failure event.}
}
\value{
A data frame with an additional column named 'event_localfailure'. This column contains
a 1 if a local failure event has occurred and a 0 otherwise.
}
\description{
This function adds a new column to the dataframe that indicates whether a local failure event
has occurred for each patient. A local failure event is defined as an occurrence of the value 2
in any of the 'overall_disease_local_recurrence_' columns.
}
\examples{
\dontrun{
# Example DataFrame (df)
# overall_disease_local_recurrence_3m overall_disease_local_recurrence_6m
#                                    0                                   2
#                                    0                                   0
#                                    2                                   0
# df <- add_local_failure_event(df)
}

}
