% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_nodalcontrol_incl_pao.R
\name{add_locoregional_event}
\alias{add_locoregional_event}
\title{Add Locoregional Event Status (Pelvic/PAO)}
\usage{
add_locoregional_event(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient event data}
}
\value{
Original dataframe with additional boolean columns 'event_locoregional' and 'event_locoregional_alone'
}
\description{
This function adds columns indicating whether a patient had a locoregional event,
defined as either a local failure or nodal control including para-aortic, and
whether the event was locoregional alone, defined as locoregional and not systemic
excluding para-aortic.
}
