% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lost_to_followup.R
\name{add_lost_to_fu}
\alias{add_lost_to_fu}
\title{Check if any follow-up is lost}
\usage{
add_lost_to_fu(df, pattern = "followup_")
}
\arguments{
\item{df}{A data frame or tibble containing follow-up columns.}

\item{pattern}{A character string specifying the pattern to identify follow-up columns. Default is `"followup"`.}
}
\value{
A data frame with an additional logical column `is_lost_to_fu` indicating if the patient is lost to follow-up.
}
\description{
This function checks whether any follow-up columns in a data frame
contain a value of `2`, which would indicate that the patient is lost to follow-up.
EMBRACE-II only
}
\examples{
df <- tibble::tibble(
  embrace_id = c("AAR2001", "VIE2001", "VIE2002"),
  followup_3m = c(1, 1, 1),
  followup_6m = c(1, 1, 1),
  followup_9m = c(1, 1, 1),
  followup_12m = c(1, -1, 1),
  followup_18m = c(1, NA, 1),
  followup_24m = c(NA, NA, -1)
)
add_lost_to_fu(df)
}
