% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_number_paraaortic_ln_statd.R
\name{add_number_paraaortic_ln_stat_d}
\alias{add_number_paraaortic_ln_stat_d}
\title{Add Number of Para Aortic Lymph Nodes per Patient}
\usage{
add_number_paraaortic_ln_stat_d(df)
}
\description{
Add Number of Para Aortic Lymph Nodes per Patient
}
