% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icis.R
\name{add_parallel_oblique_needles}
\alias{add_parallel_oblique_needles}
\title{Add parallel or oblique classification for icis}
\usage{
add_parallel_oblique_needles(.data)
}
\description{
Add parallel or oblique classification for icis
}
