% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parametrial_involvment.R
\name{add_parametrial_involvement}
\alias{add_parametrial_involvement}
\title{Add new columns to a data frame that contain the maximum parametrial involvement}
\usage{
add_parametrial_involvement(data)
}
\arguments{
\item{data}{A data frame containing at least two sets of numeric columns representing parametrial involvement on the left and right side.}
}
\value{
A data frame with new columns added.
}
\description{
This function adds several new columns to a given data frame, each containing the
maximum value between two existing columns related to parametrial involvement.
}
