% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_log.R
\name{apply_change_log}
\alias{apply_change_log}
\title{Apply Changes from the Change Log to the Dataframe}
\usage{
apply_change_log(data)
}
\arguments{
\item{data}{A dataframe containing the master data.}
}
\value{
A dataframe with the applied changes.
}
\description{
This function reads a change log from an Excel file and applies the changes to a dataframe.
It ensures the new values are correctly converted to the appropriate data types before applying the changes.
}
\details{
The change log should be an Excel file located at "data_raw/change_log/data_change_log.xlsx" with columns: `RecordID`, `Field`, `NewValue`.
The function reads the change log, converts the `NewValue` to the appropriate type based on the column type in the master dataframe, and applies the changes.
}
