% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latest_side_effects.R
\name{clean_eortc_names}
\alias{clean_eortc_names}
\title{Clean up eortc column names}
\usage{
clean_eortc_names(df, time_suffix)
}
\arguments{
\item{df}{The data frame containing the side effect columns.}

\item{time_suffix}{The suffix indicating the time point (e.g., "_3m").}
}
\value{
A vector of cleaned-up side effect column names.
}
\description{
Clean up eortc column names
}
