% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{createStyledExcelSheet}
\alias{createStyledExcelSheet}
\title{Create Styled Excel Sheet}
\usage{
createStyledExcelSheet(data, filePath, sheetName = "Side Effects")
}
\arguments{
\item{data}{DataFrame to be written to the Excel sheet.}

\item{filePath}{Character, path where the Excel file will be saved.}

\item{sheetName}{Character, name of the sheet within the Excel file.}
}
\description{
Creates an Excel sheet with styled cells based on the data provided.
}
