% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_disease_control.R
\name{emi_add_disease_control}
\alias{emi_add_disease_control}
\title{Add Disease Control Status for Embrace I}
\usage{
emi_add_disease_control(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient disease control data}
}
\value{
Original dataframe with an additional boolean column 'event_disease_control'
}
\description{
This function adds a column indicating disease control events for Embrace I data,
defined as either local failure, regional events, or systemic events.
}
\keyword{internal}
