% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_cancer_specific.R
\name{emii_add_cancer_specific}
\alias{emii_add_cancer_specific}
\title{Add Cancer-Specific Survival Status}
\usage{
emii_add_cancer_specific(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient vital status data}
}
\value{
Original dataframe with an additional boolean column 'event_cancer_specific'
}
\description{
This function adds a column indicating cancer-specific death events,
defined as death (Vital Status = 1) where the main cause of death
was primary cancer (codes 1 or 3).
}
