% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_distant_alone.R
\name{emii_add_distant_alone}
\alias{emii_add_distant_alone}
\title{Add Distant Alone Status}
\usage{
emii_add_distant_alone(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient failure data}
}
\value{
Original dataframe with an additional boolean column 'distant_alone'
}
\description{
This function adds a column indicating whether a patient had distant metastases alone,
defined as having systemic failure without para-aortic nodes and no local or nodal failures.
}
