% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_paraaortic_nodal.R
\name{emii_add_paraaortic_nodal}
\alias{emii_add_paraaortic_nodal}
\title{Add Para-aortic Nodal Control Status}
\usage{
emii_add_paraaortic_nodal(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient failure data}
}
\value{
Original dataframe with an additional boolean column 'event_paraaortic_nodal'
}
\description{
This function adds a column indicating whether a patient had para-aortic nodal involvement,
defined as either having para-aortic metastases or para-aortic nodal failure during follow-up.
}
\keyword{internal}
