% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_pelvic_event.R
\name{emii_add_pelvic_event}
\alias{emii_add_pelvic_event}
\title{Add Composite Pelvic Event Status}
\usage{
emii_add_pelvic_event(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient event data}
}
\value{
Original dataframe with an additional boolean column 'event_pelvic'
}
\description{
This function adds a column indicating whether a patient had any pelvic event,
defined as either a local failure or a pelvic nodal event.
}
\keyword{internal}
