% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_pelvic_nodal_event.R
\name{emii_add_pelvic_nodal_event}
\alias{emii_add_pelvic_nodal_event}
\title{Add Pelvic Nodal Event Status}
\usage{
emii_add_pelvic_nodal_event(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient node data}
}
\value{
Original dataframe with an additional boolean column 'event_pelvic_nodal'
}
\description{
This function adds a column indicating whether a patient had a pelvic nodal event.
A pelvic nodal event is defined as having event_nodalfailure AND at least one positive
node in any of the following locations: external iliac, internal iliac, common iliac,
or parametrial/paracervical regions (left or right).
}
\keyword{internal}
