% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_progression_free_survival.R
\name{emii_add_progression_free_survival}
\alias{emii_add_progression_free_survival}
\title{Add Progression-Free Survival Status}
\usage{
emii_add_progression_free_survival(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient event data}
}
\value{
Original dataframe with an additional boolean column 'event_progression_free'
}
\description{
This function adds a column indicating progression-free survival events,
defined as either any disease event (local failure, nodal control including PAO,
or systemic excluding PAO) OR death.
}
\keyword{internal}
