% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_systemic_excl_pao.R
\name{emii_add_systemic_excl_pao}
\alias{emii_add_systemic_excl_pao}
\title{Add Systemic Control Status (Excluding Para-aortic)}
\usage{
emii_add_systemic_excl_pao(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient failure data}
}
\value{
Original dataframe with an additional boolean column 'event_systemic_excl_pao'
}
\description{
This function adds a column indicating whether a patient had systemic failure,
defined as having systemic failure AND at least one of the following:
supraclavicular nodes, mediastinal nodes, liver metastases, bone metastases,
brain metastases, lung metastases, abdominal carcinomatosis, or other metastases.
}
\keyword{internal}
