% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_get_remission_data.R
\name{emii_get_remission_data}
\alias{emii_get_remission_data}
\title{Extract First Remission Timepoint Data from EMBRACE-II Follow-up}
\usage{
emii_get_remission_data(data)
}
\arguments{
\item{data}{A tibble containing EMBRACE-II follow-up data}
}
\value{
A tibble with remission timepoint data (embrace_id, timepoint, and all MRI measurements)
}
\description{
This function identifies the first timepoint where a patient achieves remission
(mri_remission_status_primary == 1) and extracts corresponding MRI measurements
from other variables at that timepoint.
}
\examples{
\dontrun{
  followup_data <- load_embrace_ii_followup()
  remission_data <- emii_get_remission_data(followup_data)
}
}
