% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{filter_grade_threshold}
\alias{filter_grade_threshold}
\title{Filter Data Based on Grade Threshold}
\usage{
filter_grade_threshold(df, side_effect_cols, grade_threshold = 4)
}
\arguments{
\item{df}{DataFrame containing the data to be filtered.}

\item{side_effect_cols}{Character vector specifying the columns to check against the threshold.}
}
\value{
DataFrame filtered based on the grade threshold.
}
\description{
Filters a dataframe based on specified side effect columns and a grade threshold,
retaining rows that meet the threshold criteria.
}
