% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{generatePatientExcelReports}
\alias{generatePatientExcelReports}
\title{Generate Patient Excel Reports}
\usage{
generatePatientExcelReports(df_filtered, root_folder)
}
\arguments{
\item{df_filtered}{DataFrame containing filtered patient data}

\item{root_folder}{Character, the root directory for saving reports}
}
\description{
Creates individual Excel reports for each patient in the filtered dataset,
organizing files by center ID in appropriate subfolders.
}
