% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_of_two_columns.R
\name{get_max_of_two_columns}
\alias{get_max_of_two_columns}
\title{Get the maximum of two numeric columns for each observation.}
\usage{
get_max_of_two_columns(df, col1, col2)
}
\arguments{
\item{df}{A data frame with at least two numeric columns.}

\item{col1}{The name of the first numeric column.}

\item{col2}{The name of the second numeric column.}
}
\value{
A numeric vector containing the maximum value for each observation.
}
\description{
This function computes the maximum of two numeric columns for each observation.
NA values are handled such that if one value is NA and the other is not, the non-NA value is returned.
If both values are NA, NA is returned.
}
\keyword{internal}
