% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_get_lymph_nodes_statd.R
\name{get_nodal_location_summary}
\alias{get_nodal_location_summary}
\title{Generate Summary Table of Nodal Anatomical Locations}
\usage{
get_nodal_location_summary(df)
}
\arguments{
\item{df}{A dataframe containing the EMBRACE-II dataset}
}
\value{
A gt table object showing the distribution of lymph node locations
}
\description{
Creates a formatted summary table using gtsummary showing the frequency
and percentage of each anatomical location in the nodal dataset from EMBRACE-II
at diagnosis.
}
