% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_recurrent_nodes.R
\name{get_recurrence_and_metastases_summary}
\alias{get_recurrence_and_metastases_summary}
\title{Generate Patient-Level Summary Table of Recurrent Nodes and Metastases}
\usage{
get_recurrence_and_metastases_summary(df)
}
\arguments{
\item{df}{A dataframe containing the EMBRACE-II dataset}
}
\value{
A gt table object showing the distribution of patients with recurrent nodes
  and metastases by location
}
\description{
Creates a formatted summary table using gtsummary showing the number and percentage
of patients who have recurrent lymph nodes or metastases at follow-up from EMBRACE-II.
}
