% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nodal_positions.R
\name{get_valid_locations}
\alias{get_valid_locations}
\title{Get Valid Node Locations}
\usage{
get_valid_locations()
}
\value{
Character vector of valid anatomical locations
}
\description{
Returns a vector of valid anatomical locations
}
