% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embrace_ii.R
\name{load_embrace_ii}
\alias{load_embrace_ii}
\title{Load and Process Embrace II Data}
\usage{
load_embrace_ii(
  file_path = here::here("data_raw/embrace_II/emii.xlsx"),
  file_path_eqd2 = here::here("data_raw/embrace_II/emii_eqd2.xlsx"),
  add_new_columns = T,
  mapping_file = here::here("data_raw/mapping_table/mapping_table.xlsx"),
  include_only_study_patients = TRUE,
  replace_minusone_with_na = TRUE
)
}
\arguments{
\item{file_path}{The path to the directory containing the primary data files.}

\item{file_path_eqd2}{The path to the EMBRACE-II EQD2 Excel file.}

\item{add_new_columns}{If TRUE, calculate and add new columns}

\item{mapping_file}{Path to the mapping table file}

\item{include_only_study_patients}{If TRUE, only return patients included in the study}
}
\value{
A tibble containing the cleaned and joined data.
}
\description{
This function loads data from EMBRACE-II website exports,
processes them, and joins them by the `embrace_id` column.
}
\examples{
\dontrun{
  df <- load_embrace_ii(file_path = "/path/to/xlsx", file_path_eqd2 = "/path/to/eqd2_xlsx")
}
}
