% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_column_names.R
\name{map_column_names}
\alias{map_column_names}
\title{Map Columns in a Dataframe Based on a Mapping Table}
\usage{
map_column_names(data, mapping, study = "emi")
}
\arguments{
\item{data}{The dataframe whose columns are to be renamed.}

\item{mapping}{A dataframe containing at least two columns:
`embrace_i_column` for the Embrace I dataset column names, and
`embrace_ii_column` for the Embrace II dataset column names.}

\item{study}{name of study}
}
\value{
A dataframe with columns renamed based on the mapping table.
}
\description{
This function renames columns in a given dataframe based on a mapping table.
The mapping table should contain the old column names and the new column names
that they should be mapped to.
}
\examples{
\dontrun{
  mapping <- data.frame(embrace_i_column = c('col1', 'col2'),
                        embrace_ii_column = c('column1', 'column2'),
                        new_combined_name = c('new_col1', 'new_col2'))
  data <- data.frame(col1 = 1:5, col2 = 6:10)
  new_data <- map_columns_based_on_mapping(data, mapping)
}
}
\keyword{internal}
