% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{move_row}
\alias{move_row}
\title{Move a Specific Row to a New Position}
\usage{
move_row(df, row_from, row_to)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{row_from}{The index of the row to move.}

\item{row_to}{The new index for the row.}
}
\value{
A dataframe with the row moved to the new position.
}
\description{
This function moves a specified row in a dataframe to a new position.
}
\examples{
\dontrun{
df <- tibble::tibble(
  a = 1:5,
  b = letters[1:5]
)
new_df <- move_row(df, 5, 2)
print(new_df)
}
}
\keyword{internal}
