% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nodal_positions.R
\docType{data}
\name{NODAL_POSITION_MAPPING}
\alias{NODAL_POSITION_MAPPING}
\title{Node Position Mapping Constants}
\format{
An object of class \code{character} of length 13.
}
\usage{
NODAL_POSITION_MAPPING
}
\description{
Defines the standard mapping between numeric codes and anatomical positions
for lymph nodes in the EMBRACE studies.
}
\keyword{datasets}
