% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_emii_data.R
\name{process_emii_data}
\alias{process_emii_data}
\title{Process and Transform EM-II Data}
\usage{
process_emii_data(
  data,
  file_path = here::here("data_raw/embrace_II/2023-01-18_emii_masterdump.Rds")
)
}
\arguments{
\item{data}{A tibble or dataframe containing the combined data to be processed.}

\item{file_path}{Path to the EMBRACE-II master dump file}
}
\value{
A tibble or dataframe
}
\description{
Preprocessing for EMBRACE-II. This function adds EMBRACE-II specific columns
after the common columns have been added via process_combined_data.
}
\examples{
\dontrun{
  processed_data <- process_emii_data(combined_data)
}
}
