% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_and_convert_all_columns.R
\name{recode_and_convert_all_columns}
\alias{recode_and_convert_all_columns}
\title{Recode and convert columns based on the mapping}
\usage{
recode_and_convert_all_columns(df, stop_on_missing_col = FALSE)
}
\description{
Recode and convert columns based on the mapping
}
