% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_and_convert_all_columns.R
\name{recode_and_convert_column}
\alias{recode_and_convert_column}
\title{Recode and convert columns based on the mapping}
\usage{
recode_and_convert_column(df, col_name, mapping)
}
\arguments{
\item{df}{A data frame containing the columns to be recoded and converted.}

\item{mapping}{A list containing the old and new levels for each column or base name.}

\item{stop_on_missing_col}{A logical value. If TRUE, the function stops if a column is missing in the data frame. If FALSE, it skips the missing columns.}
}
\value{
The data frame with the specified columns recoded and converted to factors.
}
\description{
This function converts columns to factors and recodes their levels based on the provided mapping.
It can process both specific column names and columns with common base names.
}
