% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_infinite_with_NA}
\alias{replace_infinite_with_NA}
\title{Replace Infinite values in numeric columns with NA}
\usage{
replace_infinite_with_NA(df)
}
\arguments{
\item{df}{The data frame containing numeric columns.}
}
\value{
A data frame where Infinite values in numeric columns are replaced with NA.
}
\description{
Replace Infinite values in numeric columns with NA
}
