% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_neg_one_with_NA}
\alias{replace_neg_one_with_NA}
\title{Replace -1 values in numeric columns with NA}
\usage{
replace_neg_one_with_NA(df)
}
\arguments{
\item{df}{The data frame containing numeric columns.}
}
\value{
A data frame where -1 values in numeric columns are replaced with NA.
}
\description{
Replace -1 values in numeric columns with NA
}
\keyword{internal}
