% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{reshape_side_effects}
\alias{reshape_side_effects}
\title{Reshape Side Effects Data}
\usage{
reshape_side_effects(df)
}
\arguments{
\item{df}{DataFrame, the data to be reshaped.}
}
\value{
DataFrame in long format with separated side effect names and time points.
}
\description{
Converts the side effects data from wide to long format, separating side effect names and time points.
}
\keyword{internal}
