% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_embrace_data.R
\name{save_embrace_data}
\alias{save_embrace_data}
\title{Save Dataframe to RDS and XLSX Formats}
\usage{
save_embrace_data(df)
}
\arguments{
\item{df}{The dataframe to be saved.}
}
\description{
This function saves a given dataframe to both RDS and XLSX formats.
The files are saved in a folder named 'data', and the filenames include the current date.
}
\examples{
\dontrun{
  data <- data.frame(x = 1:5, y = 6:10)
  save_embrace_data(data)
}
}
\keyword{internal}
