% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embrace_ii.R
\name{transpose_sheet_to_wide}
\alias{transpose_sheet_to_wide}
\title{Transpose a long-format sheet to wide-format}
\usage{
transpose_sheet_to_wide(file_path, sheet_name, id_col, key_col)
}
\arguments{
\item{file_path}{The path to the Excel file.}

\item{sheet_name}{The name of the sheet to import.}

\item{id_col}{The column name that contains unique identifiers.}

\item{key_col}{The column name that will become the new columns in the wide format.}
}
\value{
A tibble in wide format.
}
\description{
Transpose a long-format sheet to wide-format
}
\keyword{internal}
