% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hrctv_volume_bins.R
\name{add_hrctv_volume_bins}
\alias{add_hrctv_volume_bins}
\title{Bins values in a specified column of a dataframe}
\usage{
add_hrctv_volume_bins(df)
}
\arguments{
\item{df}{A dataframe containing the volume column to be binned.}
}
\value{
A dataframe with the original data and a new column 'fraction01hrctv_volume_bins' with encoded values.
}
\description{
Takes a dataframe and bins the values of the specified column into three categories:
less than 30, between 30 and 45 inclusive, and greater than 45. These bins are
then encoded as 0, 1, and 2, respectively. Missing values are left as NA.
}
\examples{
# Example usage:
library(dplyr)
sample_df <- data.frame(fraction01hrctv_volume_tdvh = c(10, 35, 50, 22, 40, NA))
binned_df <- add_hrctv_volume_bins(sample_df)
print(binned_df)
}
