% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_to_diseaseevent.R
\name{add_time_to_diseaseevent}
\alias{add_time_to_diseaseevent}
\title{Calculate the latest assessment date and follow-up time to the disease event}
\usage{
add_time_to_diseaseevent(df)
}
\arguments{
\item{df}{A data frame with columns that include assessment dates, histology assessment dates,
and disease recurrence information.}
}
\value{
A data frame with added columns for the latest assessment date and the time to the event in months.
}
\description{
This function processes the dataframe to find the latest assessment date before a disease event.
If a disease event (as defined in the corresponding columns) occurs, this date is considered the latest.
}
