% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_to_vitalstatus.R
\name{add_time_to_last_vitalstatus}
\alias{add_time_to_last_vitalstatus}
\title{Calculate Time to Last Vital Status in Months}
\usage{
add_time_to_last_vitalstatus(df)
}
\arguments{
\item{df}{A data frame that contains the required columns based on study type}
}
\value{
A data frame with the new or updated `timetoevent_vitalstatus` column.
}
\description{
This function adds a new column named `timetoevent_vitalstatus` to the given data frame.
For EMBRACE I, it calculates the time interval in months between `histology_assesment_date`
and the latest of `vital_status_last_info_date` or assessment dates.
For EMBRACE II, it also considers `vital_status_date_of_death_vital_status` when available.
}
\examples{
\dontrun{
# Given a DataFrame df with the required columns
# df <- add_time_to_last_vitalstatus(df)
}

}
