% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_recurrent_nodes.R
\name{create_combined_labels}
\alias{create_combined_labels}
\title{Helper function to create combined labels for gtsummary}
\usage{
create_combined_labels()
}
\description{
Helper function to create combined labels for gtsummary
}
\keyword{internal}
