% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_recurrent_nodes.R
\name{emii_add_diagnostic_nodes}
\alias{emii_add_diagnostic_nodes}
\title{Add Diagnostic Node Locations to EMII Data}
\usage{
emii_add_diagnostic_nodes(.data)
}
\arguments{
\item{.data}{The input dataframe containing patient node data}
}
\value{
Original dataframe with additional boolean columns for each nodal location at diagnosis
}
\description{
This function adds columns for all possible nodal locations at diagnosis to the input dataset.
}
\examples{
\dontrun{
patient_data <- emii_add_diagnostic_nodes(patient_data)
}
}
