% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_add_number_comiliac_ln_statd.R
\name{emii_add_number_common_iliac_ln_stat_d}
\alias{emii_add_number_common_iliac_ln_stat_d}
\title{Add Number of Common Iliac Lymph Nodes per Patient}
\usage{
emii_add_number_common_iliac_ln_stat_d(df)
}
\description{
Add Number of Common Iliac Lymph Nodes per Patient
}
