% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_get_critical_parameter_status.R
\name{emii_get_critical_parameter_status}
\alias{emii_get_critical_parameter_status}
\title{Get Status of Critical Parameters in a Dataframe}
\usage{
emii_get_critical_parameter_status(save_excel = FALSE)
}
\value{
A gtsummary object containing the combined summary table.
}
\description{
This function generates a summary table comparing critical parameters between EMBRACE I and II datasets.
It also calculates the proportion of missing data for each parameter. Optionally, the summary
table can be saved as an Excel file.
}
