% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emii_get_dosimetric_parameter_status.R
\name{emii_get_dosimetric_parameter_status}
\alias{emii_get_dosimetric_parameter_status}
\title{Get Status of Dosimetric Parameters and Compare with EMBRACE-I}
\usage{
emii_get_dosimetric_parameter_status()
}
\value{
A gtsummary object containing the comparative summary table.
}
\description{
Get Status of Dosimetric Parameters and Compare with EMBRACE-I
}
