% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{generateExcelReportsForPatients}
\alias{generateExcelReportsForPatients}
\title{Generate Excel Reports for Patients}
\usage{
generateExcelReportsForPatients(data, outputDir, grade_threshold)
}
\arguments{
\item{data}{DataFrame, the data to be processed.}

\item{outputDir}{Character, the output directory for the Excel reports.}

\item{grade_threshold}{Grade threshold}
}
\description{
Cleans side effect names, filters the data based on a grade threshold, and generates Excel reports for each patient.
}
