% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{generatePatientExcelReports}
\alias{generatePatientExcelReports}
\title{Generate Patient Excel Reports}
\usage{
generatePatientExcelReports(df_filtered, root_folder)
}
\arguments{
\item{df_filtered}{DataFrame, the filtered data for which reports will be generated.}

\item{root_folder}{Character, the root folder where subfolders and Excel files will be saved.}
}
\description{
For each row in the filtered dataframe, generates an Excel report with a filename based on the patient ID.
}
