% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_combined_embrace.R
\name{load_combined_embrace}
\alias{load_combined_embrace}
\title{Load and Combine Embrace I and Embrace II Data}
\usage{
load_combined_embrace(
  file_path_i = here::here("data_raw/embrace_I"),
  file_path_ii = here::here("data_raw/embrace_II/emii.xlsx"),
  file_path_eqd2 = here::here("data_raw/embrace_II/emii_eqd2.xlsx"),
  mapping_file = here::here("data_raw/mapping_table/mapping_table.xlsx"),
  return_common_columns = T
)
}
\arguments{
\item{file_path_i}{The path to the EMBRACE-I data root folder.}

\item{file_path_ii}{The path to the EMBRACE-II Excel file.}

\item{file_path_eqd2}{The path to the EMBRACE-II EQD2 Excel file.}

\item{return_common_columns}{If TRUE, return only matched columns}
}
\value{
A tibble containing the cleaned and combined data from both studies.
}
\description{
This function loads data from EMBRACE-I and EMBRACE-II, processes them,
and combines them into a single dataframe.
}
