% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embrace_i.R
\name{load_embrace_i}
\alias{load_embrace_i}
\title{Load and Process Embrace I Data}
\usage{
load_embrace_i(
  file_path = here::here("data_raw/embrace_I"),
  add_new_columns = T,
  mapping_file = here::here("data_raw/mapping_table/mapping_table.xlsx"),
  filter_cohort = T
)
}
\arguments{
\item{file_path}{The path to the directory containing the primary data files.}

\item{add_new_columns}{If TRUE, calculate and add new columns}

\item{mapping_file}{Path to the mapping table file}

\item{filter_cohort}{Filter 1318}
}
\value{
A tibble containing the cleaned and joined data.
}
\description{
This function loads data from two different sources for the Embrace I study,
processes them, and joins them by the `embrace_id` column.
}
\examples{
\dontrun{
  df <- load_embrace_i()
}
}
