% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_combined_data.R
\name{process_combined_data}
\alias{process_combined_data}
\title{Process and Transform Combined Data}
\usage{
process_combined_data(data)
}
\arguments{
\item{data}{A tibble or dataframe containing the combined data to be processed.}
}
\value{
A tibble or dataframe
}
\description{
This function takes a combined dataset and applies transformations
}
\examples{
\dontrun{
  processed_data <- process_combined_data(combined_data)
}
}
