% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_emi_data.R
\name{process_emi_data}
\alias{process_emi_data}
\title{Process and Transform EM-I Data}
\usage{
process_emi_data(data)
}
\arguments{
\item{data}{A tibble or dataframe containing the combined data to be processed.}
}
\value{
A tibble or dataframe
}
\description{
Preprocessing for EMBRACE-I
}
\examples{
\dontrun{
  processed_data <- process_emi_data(combined_data)
}
}
