% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_and_convert_all_columns.R
\name{recode_and_convert_column}
\alias{recode_and_convert_column}
\title{Recode and convert a single column based on the mapping}
\usage{
recode_and_convert_column(df, col_name, mapping)
}
\arguments{
\item{df}{A data frame containing the column to be recoded}

\item{col_name}{Character, name of the column to recode}

\item{mapping}{A list containing the old and new levels for the column}
}
\value{
The data frame with the specified column recoded and converted to factor
}
\description{
Recode and convert a single column based on the mapping
}
\keyword{internal}
