% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{reorder_columns_based_on_date}
\alias{reorder_columns_based_on_date}
\title{Reorder Columns Based on Assessment Dates}
\usage{
reorder_columns_based_on_date(df)
}
\arguments{
\item{df}{DataFrame with columns to be reordered}
}
\value{
DataFrame with columns reordered chronologically
}
\description{
Reorders columns in a dataframe chronologically based on their corresponding 
assessment dates, facilitating temporal analysis of side effects.
}
\keyword{internal}
