% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_morbidity_events.R
\name{threshold}
\alias{threshold}
\title{Apply Threshold to Data}
\usage{
threshold(x, grade_threshold)
}
\arguments{
\item{x}{Numeric, the value to be checked against the threshold.}

\item{grade_threshold}{Numeric, the threshold value to compare against.}
}
\value{
Logical, TRUE if the value is above the threshold, otherwise FALSE.
}
\description{
This function applies a threshold to a given data set to determine if the grade
is above a specified level.
}
\examples{
threshold(5, grade_threshold = 4)
threshold(NA, grade_threshold = 4)
}
